refactor(frontend):重构龙虎榜功能
- 将龙虎榜相关代码从 market.vue 中抽离，创建独立的 LongTigerRankList 组件
-优化龙虎榜数据获取逻辑，增加对历史数据的递归查询
- 改进用户界面，保留原有的筛选和排序功能
- 删除 market.vue 中的冗余代码，提高代码可读性和维护性
